requirements.txt
requests
chromadb
python-dotenv
pyTelegramBotAPI
tavily-python
edge-tts
langdetect
pygame
networkx

SpeechRecognition
pyaudio
